More geometry generator clean up.
