+++ editor/examples/camera.app.json
@@ -37,7 +37,7 @@
 		"geometries": [
 			{
 				"uuid": "6D90C4BE-EBA6-4E21-8F54-7CFDAA61F30B",
-				"type": "PlaneBufferGeometry",
+				"type": "PlaneGeometry",
 				"width": 10,
 				"height": 10,
 				"widthSegments": 1,
@@ -45,7 +45,7 @@
 			},
 			{
 				"uuid": "D3008B2A-ACDD-43CC-87F7-4F942607D21A",
-				"type": "BoxBufferGeometry",
+				"type": "BoxGeometry",
 				"width": 1,
 				"height": 1,
 				"depth": 1,
@@ -55,7 +55,7 @@
 			},
 			{
 				"uuid": "F482ACD4-013A-49CF-AE0F-C9FF4ADAE409",
-				"type": "CylinderBufferGeometry",
+				"type": "CylinderGeometry",
 				"radiusTop": 0,
 				"radiusBottom": 0.4,
 				"height": 0.75,
@@ -65,7 +65,7 @@
 			},
 			{
 				"uuid": "51CDDCED-BC71-4B1B-A485-725B6A48204B",
-				"type": "IcosahedronBufferGeometry",
+				"type": "IcosahedronGeometry",
 				"radius": 0.4,
 				"detail": 2
 			}],
