+++ editor/examples/particles.app.json
@@ -37,7 +37,7 @@
 		"geometries": [
 			{
 				"uuid": "C3C0CE7D-10B8-43FC-8F74-011CC6E57800",
-				"type": "PlaneBufferGeometry",
+				"type": "PlaneGeometry",
 				"width": 100,
 				"height": 100,
 				"widthSegments": 1,
