+++ editor/examples/pong.app.json
@@ -37,7 +37,7 @@
 		"geometries": [
 			{
 				"uuid": "490CEBA3-6A25-4BE1-B517-C5FB11A5D18A",
-				"type": "PlaneBufferGeometry",
+				"type": "PlaneGeometry",
 				"width": 6,
 				"height": 4,
 				"widthSegments": 1,
@@ -45,7 +45,7 @@
 			},
 			{
 				"uuid": "D9A92F2D-2F08-4851-99C7-12D8D1CA13C7",
-				"type": "BoxBufferGeometry",
+				"type": "BoxGeometry",
 				"width": 0.1,
 				"height": 0.1,
 				"depth": 0.1,
@@ -55,7 +55,7 @@
 			},
 			{
 				"uuid": "5E63B8CF-E225-4ABC-994A-4D06BD4E21EB",
-				"type": "BoxBufferGeometry",
+				"type": "BoxGeometry",
 				"width": 0.2,
 				"height": 0.2,
 				"depth": 1,
@@ -65,7 +65,7 @@
 			},
 			{
 				"uuid": "D61532B4-24C3-4BC4-B56B-7245E8163E09",
-				"type": "BoxBufferGeometry",
+				"type": "BoxGeometry",
 				"width": 0.2,
 				"height": 0.2,
 				"depth": 1,
