+++ editor/examples/shaders.app.json
@@ -37,7 +37,7 @@
 		"geometries": [
 			{
 				"uuid": "EA781333-F3AE-470D-9110-A9724FCB42AA",
-				"type": "IcosahedronBufferGeometry",
+				"type": "IcosahedronGeometry",
 				"radius": 1,
 				"detail": 4
 			}],
