+++ editor/js/Sidebar.Geometry.DodecahedronGeometry.js
@@ -8,8 +8,6 @@ function GeometryParametersPanel( editor, object ) {
 
 	var strings = editor.strings;
 
-	var signals = editor.signals;
-
 	var container = new UIRow();
 
 	var geometry = object.geometry;
@@ -20,7 +18,7 @@ function GeometryParametersPanel( editor, object ) {
 	var radiusRow = new UIRow();
 	var radius = new UINumber( parameters.radius ).onChange( update );
 
-	radiusRow.add( new UIText( strings.getKey( 'sidebar/geometry/icosahedron_geometry/radius' ) ).setWidth( '90px' ) );
+	radiusRow.add( new UIText( strings.getKey( 'sidebar/geometry/dodecahedron_geometry/radius' ) ).setWidth( '90px' ) );
 	radiusRow.add( radius );
 
 	container.add( radiusRow );
@@ -30,7 +28,7 @@ function GeometryParametersPanel( editor, object ) {
 	var detailRow = new UIRow();
 	var detail = new UIInteger( parameters.detail ).setRange( 0, Infinity ).onChange( update );
 
-	detailRow.add( new UIText( strings.getKey( 'sidebar/geometry/icosahedron_geometry/detail' ) ).setWidth( '90px' ) );
+	detailRow.add( new UIText( strings.getKey( 'sidebar/geometry/dodecahedron_geometry/detail' ) ).setWidth( '90px' ) );
 	detailRow.add( detail );
 
 	container.add( detailRow );
@@ -39,12 +37,10 @@ function GeometryParametersPanel( editor, object ) {
 
 	function update() {
 
-		editor.execute( new SetGeometryCommand( editor, object, new THREE.IcosahedronBufferGeometry(
+		editor.execute( new SetGeometryCommand( editor, object, new THREE.DodecahedronGeometry(
 			radius.getValue(),
 			detail.getValue()
 		) ) );
-
-		signals.objectChanged.dispatch( object );
 
 	}
