+++ editor/sw.js
@@ -61,7 +61,7 @@ const assets = [
 
 	'../examples/jsm/helpers/VertexNormalsHelper.js',
 
-	'../examples/jsm/geometries/TeapotBufferGeometry.js',
+	'../examples/jsm/geometries/TeapotGeometry.js',
 
 	'../examples/jsm/webxr/VRButton.js',
 
@@ -140,23 +140,23 @@ const assets = [
 	'./js/Sidebar.Geometry.Geometry.js',
 	'./js/Sidebar.Geometry.BufferGeometry.js',
 	'./js/Sidebar.Geometry.Modifiers.js',
-	'./js/Sidebar.Geometry.BoxBufferGeometry.js',
-	'./js/Sidebar.Geometry.CircleBufferGeometry.js',
-	'./js/Sidebar.Geometry.CylinderBufferGeometry.js',
-	'./js/Sidebar.Geometry.DodecahedronBufferGeometry.js',
-	'./js/Sidebar.Geometry.ExtrudeBufferGeometry.js',
-	'./js/Sidebar.Geometry.IcosahedronBufferGeometry.js',
-	'./js/Sidebar.Geometry.LatheBufferGeometry.js',
-	'./js/Sidebar.Geometry.OctahedronBufferGeometry.js',
-	'./js/Sidebar.Geometry.PlaneBufferGeometry.js',
-	'./js/Sidebar.Geometry.RingBufferGeometry.js',
-	'./js/Sidebar.Geometry.SphereBufferGeometry.js',
-	'./js/Sidebar.Geometry.ShapeBufferGeometry.js',
-	'./js/Sidebar.Geometry.TetrahedronBufferGeometry.js',
-	'./js/Sidebar.Geometry.TorusBufferGeometry.js',
-	'./js/Sidebar.Geometry.TorusKnotBufferGeometry.js',
-	'./js/Sidebar.Geometry.TubeBufferGeometry.js',
-	'./js/Sidebar.Geometry.TeapotBufferGeometry.js',
+	'./js/Sidebar.Geometry.BoxGeometry.js',
+	'./js/Sidebar.Geometry.CircleGeometry.js',
+	'./js/Sidebar.Geometry.CylinderGeometry.js',
+	'./js/Sidebar.Geometry.DodecahedronGeometry.js',
+	'./js/Sidebar.Geometry.ExtrudeGeometry.js',
+	'./js/Sidebar.Geometry.IcosahedronGeometry.js',
+	'./js/Sidebar.Geometry.LatheGeometry.js',
+	'./js/Sidebar.Geometry.OctahedronGeometry.js',
+	'./js/Sidebar.Geometry.PlaneGeometry.js',
+	'./js/Sidebar.Geometry.RingGeometry.js',
+	'./js/Sidebar.Geometry.SphereGeometry.js',
+	'./js/Sidebar.Geometry.ShapeGeometry.js',
+	'./js/Sidebar.Geometry.TetrahedronGeometry.js',
+	'./js/Sidebar.Geometry.TorusGeometry.js',
+	'./js/Sidebar.Geometry.TorusKnotGeometry.js',
+	'./js/Sidebar.Geometry.TubeGeometry.js',
+	'./js/Sidebar.Geometry.TeapotGeometry.js',
 	'./js/Sidebar.Material.js',
 	'./js/Sidebar.Animation.js',
 	'./js/Sidebar.Script.js',
