+++ examples/jsm/geometries/TeapotGeometry.d.ts
@@ -2,7 +2,7 @@ import {
 	BufferGeometry
 } from '../../../src/Three';
 
-export class TeapotBufferGeometry extends BufferGeometry {
+export class TeapotGeometry extends BufferGeometry {
 
 	constructor( size?: number, segments?: number, bottom?: boolean, lid?: boolean, body?: boolean, fitLid?: boolean, blinn?: number );
