+++ examples/jsm/geometries/TeapotGeometry.js
@@ -1,3 +1,11 @@
+import {
+	BufferAttribute,
+	BufferGeometry,
+	Matrix4,
+	Vector3,
+	Vector4
+} from '../../../build/three.module.js';
+
 /**
  * Tessellates the famous Utah teapot database by Martin Newell into triangles.
  *
@@ -47,7 +55,7 @@
  *
  */
 
-THREE.TeapotBufferGeometry = function ( size, segments, bottom, lid, body, fitLid, blinn ) {
+var TeapotGeometry = function ( size, segments, bottom, lid, body, fitLid, blinn ) {
 
 	// 32 * 4 * 4 Bezier spline patches
 	var teapotPatches = [
@@ -384,7 +392,7 @@ THREE.TeapotBufferGeometry = function ( size, segments, bottom, lid, body, fitLi
 		1.5, - 0.84, 0.075
 	];
 
-	THREE.BufferGeometry.call( this );
+	BufferGeometry.call( this );
 
 	size = size || 50;
 
@@ -433,7 +441,7 @@ THREE.TeapotBufferGeometry = function ( size, segments, bottom, lid, body, fitLi
 	var uvs = new Float32Array( numVertices * 2 );
 
 	// Bezier form
-	var ms = new THREE.Matrix4();
+	var ms = new Matrix4();
 	ms.set(
 		- 1.0, 3.0, - 3.0, 1.0,
 		3.0, - 6.0, 3.0, 0.0,
@@ -456,7 +464,7 @@ THREE.TeapotBufferGeometry = function ( size, segments, bottom, lid, body, fitLi
 	var sdir = [];
 	var tdir = [];
 
-	var norm = new THREE.Vector3();
+	var norm = new Vector3();
 
 	var tcoord;
 
@@ -467,19 +475,19 @@ THREE.TeapotBufferGeometry = function ( size, segments, bottom, lid, body, fitLi
 	var dsval = 0;
 	var dtval = 0;
 
-	var normOut = new THREE.Vector3();
+	var normOut = new Vector3();
 	var v1, v2, v3, v4;
 
-	var gmx = new THREE.Matrix4();
-	var tmtx = new THREE.Matrix4();
+	var gmx = new Matrix4();
+	var tmtx = new Matrix4();
 
-	var vsp = new THREE.Vector4();
-	var vtp = new THREE.Vector4();
-	var vdsp = new THREE.Vector4();
-	var vdtp = new THREE.Vector4();
+	var vsp = new Vector4();
+	var vtp = new Vector4();
+	var vdsp = new Vector4();
+	var vdtp = new Vector4();
 
-	var vsdir = new THREE.Vector3();
-	var vtdir = new THREE.Vector3();
+	var vsdir = new Vector3();
+	var vtdir = new Vector3();
 
 	var mst = ms.clone();
 	mst.transpose();
@@ -504,7 +512,7 @@ THREE.TeapotBufferGeometry = function ( size, segments, bottom, lid, body, fitLi
 
 	for ( i = 0; i < 3; i ++ ) {
 
-		mgm[ i ] = new THREE.Matrix4();
+		mgm[ i ] = new Matrix4();
 
 	}
 
@@ -700,15 +708,17 @@ THREE.TeapotBufferGeometry = function ( size, segments, bottom, lid, body, fitLi
 
 	}
 
-	this.setIndex( new THREE.BufferAttribute( indices, 1 ) );
-	this.setAttribute( 'position', new THREE.BufferAttribute( vertices, 3 ) );
-	this.setAttribute( 'normal', new THREE.BufferAttribute( normals, 3 ) );
-	this.setAttribute( 'uv', new THREE.BufferAttribute( uvs, 2 ) );
+	this.setIndex( new BufferAttribute( indices, 1 ) );
+	this.setAttribute( 'position', new BufferAttribute( vertices, 3 ) );
+	this.setAttribute( 'normal', new BufferAttribute( normals, 3 ) );
+	this.setAttribute( 'uv', new BufferAttribute( uvs, 2 ) );
 
 	this.computeBoundingSphere();
 
 };
 
 
-THREE.TeapotBufferGeometry.prototype = Object.create( THREE.BufferGeometry.prototype );
-THREE.TeapotBufferGeometry.prototype.constructor = THREE.TeapotBufferGeometry;
+TeapotGeometry.prototype = Object.create( BufferGeometry.prototype );
+TeapotGeometry.prototype.constructor = TeapotGeometry;
+
+export { TeapotGeometry };
