+++ examples/webgl_buffergeometry_compression.html
@@ -20,7 +20,7 @@
 			import { OrbitControls } from './jsm/controls/OrbitControls.js';
 			import { GeometryCompressionUtils } from './jsm/utils/GeometryCompressionUtils.js';
 			import { BufferGeometryUtils } from './jsm/utils/BufferGeometryUtils.js';
-			import { TeapotBufferGeometry } from './jsm/geometries/TeapotBufferGeometry.js';
+			import { TeapotGeometry } from './jsm/geometries/TeapotGeometry.js';
 			import { GUI } from './jsm/libs/dat.gui.module.js';
 
 			const statsEnabled = true;
@@ -129,7 +129,7 @@
 						case "Cylinder":
 							return new THREE.CylinderGeometry( radius, radius, radius * 2, data.detail * 6 );
 						case "Teapot":
-							return new TeapotBufferGeometry( radius, data.detail * 3, true, true, true, true, true );
+							return new TeapotGeometry( radius, data.detail * 3, true, true, true, true, true );
 						case "TorusKnot":
 							return new THREE.TorusKnotGeometry( radius, 10, data.detail * 20, data.detail * 6, 3, 4 );
