+++ examples/webgl_materials_compile.html
@@ -38,7 +38,7 @@
 			import * as THREE from '../build/three.module.js';
 
 			import { OrbitControls } from './jsm/controls/OrbitControls.js';
-			import { TeapotBufferGeometry } from './jsm/geometries/TeapotBufferGeometry.js';
+			import { TeapotGeometry } from './jsm/geometries/TeapotGeometry.js';
 
 			import {
 				NodeFrame,
@@ -109,7 +109,7 @@
 				light2.position.set( - 1, 0.75, - 0.5 );
 				scene.add( light2 );
 
-				teapot = new TeapotBufferGeometry( 15, 18 );
+				teapot = new TeapotGeometry( 15, 18 );
 
 				const itemsonrow = 10;
